+++ secrets/ed25519_test.go
@@ -5,4 +5,5 @@ import "testing"
 func TestEd25519(t *testing.T) {
 	// TODO: GenerateEd25519Keys
 	// TODO: LoadEd25519PrivateKey
+	// TODO: LoadEd25519PublicKey
 }