+++ CHANGELOG.md
@@ -2,6 +2,7 @@
   * Commit: 
   * Generate Ed25519 Keys
   * Load Ed25519 Private Key from File
+  * Load Ed25519 Public Key from File
 ## v0.3.32 - Combinations
   * Commit: 2026-05-22 06:59 
   * Combinations